created fragment MainFragmentSecondary and refactored MainActivity to delegate slidingup panel to the fragment
