+++ app/src/main/java/com/example/musicplayer/MainActivity.java
@@ -1,16 +1,11 @@
 package com.example.musicplayer;
 
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.cardview.widget.CardView;
 import androidx.core.app.ActivityCompat;
 import androidx.core.app.NotificationCompat;
 import androidx.core.app.NotificationManagerCompat;
 import androidx.core.content.ContextCompat;
-import androidx.core.content.res.ResourcesCompat;
-import androidx.core.graphics.drawable.DrawableCompat;
-
 import android.Manifest;
-import android.animation.ObjectAnimator;
 import android.annotation.SuppressLint;
 import android.annotation.TargetApi;
 import android.app.Notification;
@@ -20,13 +15,9 @@ import android.content.ContentResolver;
 import android.content.ContentUris;
 import android.content.Intent;
 import android.content.pm.PackageManager;
-import android.content.res.ColorStateList;
 import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.graphics.drawable.Drawable;
-import android.graphics.drawable.GradientDrawable;
-import android.graphics.drawable.RippleDrawable;
 import android.media.AudioManager;
 import android.net.Uri;
 import android.os.Bundle;
@@ -39,31 +30,25 @@ import android.provider.MediaStore;
 import android.support.v4.media.session.MediaSessionCompat;
 import android.view.ActionMode;
 import android.view.View;
-import android.widget.Button;
-import android.widget.ImageButton;
-import android.widget.ImageView;
-import android.widget.RelativeLayout;
-import android.widget.SeekBar;
-import android.widget.TextView;
 import android.widget.Toast;
-import com.sothree.slidinguppanel.SlidingUpPanelLayout;
-
 import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.HashMap;
-import java.util.Random;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
-
 import static com.example.musicplayer.Notifications.CHANNEL_ID_1;
+
+import com.sothree.slidinguppanel.SlidingUpPanelLayout;
 
 public class MainActivity extends AppCompatActivity {
 
     private static final int MY_PERMISSION_REQUEST = 1;
     private boolean isPermissionGranted = false;
     private boolean isInfoDisplaying;
+    private static Song current_song = Song.EMPTY_SONG;
+    private Bitmap current_albumImage;
     private ArrayList<Song> fullSongList;
     private HashMap<Integer, SongMetadata> fullSongMetadataHashMap;
     private static ArrayList<Playlist> playlistList;
@@ -71,83 +56,34 @@ public class MainActivity extends AppCompatActivity {
     private static Playlist fullPlaylist;
     private Messenger mainActivityMessenger;
     private Intent musicServiceIntent;
+    private Intent notificationIntent;
+    private Intent notificationPauseplayIntent;
+    private Intent notificationPrevIntent;
+    private Intent notificationNextIntent;
+    private Intent seekBar_progressIntent;
+    private MediaSessionCompat mediaSession;
+    private static Notification notificationChannel1;
+    private NotificationManagerCompat notificationManager;
+    private NotificationCompat.Builder notificationBuilder;
     public static boolean isActionMode = false;
     public static ActionMode actionMode = null;
     private boolean isDestroyed = false;
     private Metadata metadata;
     private boolean isMetadataLoaded = false;
-
-    // sliding up panel
     private static int random_seed;
-    private boolean isAlbumArtCircular;
     private static boolean isShuffled;
+    private boolean isAlbumArtCircular;
     private static int repeat_status;
     private MessageHandler messageHandler;
-    private MessageHandler seekbarHandler;
-    private ImageView mainDisplay_albumArt;
-    private Button mainDisplay_albumArt_btn;
-    private CardView mainDisplay_albumArt_cardView;
-    private ObjectAnimator mainDisplay_albumArt_cardView_animator_round;
-    private ObjectAnimator mainDisplay_albumArt_cardView_animator_square;
-    private GradientDrawable mainDisplay_mainGradient;
-    private SeekBar mainDisplay_seekBar;
-    private ImageButton mainDisplay_slidedown_btn;
-    private ImageButton mainDisplay_info_btn;
-    private ImageButton mainDisplay_pauseplay_btn;
-    private ImageButton mainDisplay_next_btn;
-    private ImageButton mainDisplay_prev_btn;
-    private ImageButton mainDisplay_shuffle_btn;
-    private ImageButton mainDisplay_repeat_btn;
-    private RippleDrawable mainDisplay_slidedown_btn_ripple;
-    private RippleDrawable mainDisplay_info_btn_ripple;
-    private RippleDrawable mainDisplay_prev_btn_ripple;
-    private RippleDrawable mainDisplay_pauseplay_btn_ripple;
-    private RippleDrawable mainDisplay_next_btn_ripple;
-    private RippleDrawable mainDisplay_shuffle_btn_ripple;
-    private RippleDrawable mainDisplay_repeat_btn_ripple;
-    private TextView mainDisplay_playlistHeader;
-    private TextView mainDisplay_musicPosition;
-    private TextView mainDisplay_musicDuration;
-    private TextView mainDisplay_songTitle;
-    private TextView mainDisplay_songArtist;
-    private MediaSessionCompat mediaSession;
-    private NotificationManagerCompat notificationManager;
-    private NotificationCompat.Builder notificationBuilder;
-    private static Notification notificationChannel1;
-    private Intent mainPausePlayIntent;
-    private Intent mainPrevIntent;
-    private Intent mainNextIntent;
-    private Intent infoIntent;
-    private Intent notificationPauseplayIntent;
-    private Intent notificationPrevIntent;
-    private Intent notificationNextIntent;
-    private Intent seekBar_progressIntent;
-    private Intent seekBar_seekIntent;
-    private boolean seekBar_isTracking;
-    private static Song current_song = Song.EMPTY_SONG;
-    private Bitmap current_albumImage;
-    private RelativeLayout mainActivityRelativeLayout;
-    private RelativeLayout slidingUpMenuLayout;
-    private SlidingUpPanelLayout slidingUpPanelLayout;
-    private ImageView slidingUp_albumArt;
-    private TextView slidingUp_songName;
-    private TextView slidingUp_artistName;
-    private Intent slidingUp_pauseplayIntent;
-    private Intent slidingUp_prevIntent;
-    private Intent slidingUp_nextIntent;
-    private Intent notificationIntent;
-    private ImageButton slidingUp_prev_btn;
-    private ImageButton slidingUp_pauseplay_btn;
-    private ImageButton slidingUp_next_btn;
-    private RippleDrawable slidingUp_prev_btn_ripple;
-    private RippleDrawable slidingUp_pauseplay_btn_ripple;
-    private RippleDrawable slidingUp_next_btn_ripple;
+    private MainActivity.MessageHandler seekbarHandler;
+    private SlidingUpPanelLayout mainActivityLayout;
 
     // database
     private DatabaseRepository databaseRepository;
 
     //fragments
-    private MainFragment mainFragment;
+    private MainFragmentPrimary mainFragmentPrimary;
+    private MainFragmentSecondary mainFragmentSecondary;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -167,42 +103,28 @@ public class MainActivity extends AppCompatActivity {
             messageHandler = new MessageHandler(messageHandlerThread.getLooper());
             mainActivityMessenger = new Messenger(messageHandler);
 
+            // init thread for seekbar updates
+            HandlerThread seekbarHandlerThread = new HandlerThread("SeekbarHandler");
+            seekbarHandlerThread.start();
+            seekbarHandler = new MainActivity.MessageHandler(seekbarHandlerThread.getLooper());
+
+            // init seekbar intent
+            seekBar_progressIntent = new Intent(this, MusicPlayerService.class);
+
             // when a configuration change occurs and activity is recreated, fragment is auto restored
             if (savedInstanceState == null) {
-                mainFragment = MainFragment.getInstance(mainActivityMessenger);
+                mainFragmentPrimary = MainFragmentPrimary.getInstance(mainActivityMessenger);
+                mainFragmentSecondary = MainFragmentSecondary.getInstance(mainActivityMessenger);
                 getSupportFragmentManager().beginTransaction()
                         .setReorderingAllowed(true)
-                        .add(R.id.fragment_playlist, mainFragment)
+                        .add(R.id.fragment_main_primary, mainFragmentPrimary)
+                        .add(R.id.fragment_main_secondary, mainFragmentSecondary)
                         .commit();
             }
 
-            // initialize all views
             setContentView(R.layout.activity_main);
-            slidingUpMenuLayout = findViewById(R.id.sliding_menu);
-            slidingUp_albumArt = findViewById(R.id.sliding_albumart);
-            slidingUp_songName = findViewById(R.id.sliding_title);
-            slidingUp_artistName = findViewById(R.id.sliding_artist);
-            slidingUpPanelLayout = findViewById(R.id.slidingPanel);
-            slidingUp_pauseplay_btn = findViewById(R.id.sliding_btn_play);
-            slidingUp_next_btn = findViewById(R.id.sliding_btn_next);
-            slidingUp_prev_btn = findViewById(R.id.sliding_btn_prev);
-            mainActivityRelativeLayout = findViewById(R.id.mainlayout);
-            mainDisplay_albumArt = findViewById(R.id.song_albumart);
-            mainDisplay_albumArt_cardView = findViewById(R.id.song_cardview);
-            mainDisplay_albumArt_btn = findViewById(R.id.toggle_largeAlbumArt);
-            mainDisplay_songTitle = findViewById(R.id.song_title);
-            mainDisplay_songArtist = findViewById(R.id.song_artist);
-            mainDisplay_seekBar = findViewById(R.id.seekBar);
-            mainDisplay_slidedown_btn = findViewById(R.id.slidedown_btn);
-            mainDisplay_info_btn = findViewById(R.id.btn_info);
-            mainDisplay_pauseplay_btn = findViewById(R.id.btn_play);
-            mainDisplay_next_btn = findViewById(R.id.btn_next);
-            mainDisplay_prev_btn = findViewById(R.id.btn_prev);
-            mainDisplay_shuffle_btn = findViewById(R.id.btn_shuffle);
-            mainDisplay_repeat_btn = findViewById(R.id.btn_repeat);
-            mainDisplay_musicPosition = findViewById(R.id.music_position);
-            mainDisplay_musicDuration = findViewById(R.id.music_duration);
-            mainDisplay_playlistHeader = findViewById(R.id.playlist_header);
+            mainActivityLayout = findViewById(R.id.slidingPanel);
+
         }catch (Exception e){
             Logger.logException(e);
         }
@@ -228,10 +150,6 @@ public class MainActivity extends AppCompatActivity {
 
                 // init listview functionality and playlist
                 initMusicList();
-
-                initMainGradient();
-                initMainButtons();
-                initInfoButton();
             } else if (ActivityCompat.shouldShowRequestPermissionRationale(MainActivity.this,
                     Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
                 // present rationale to user and then request for permissions
@@ -422,7 +340,7 @@ public class MainActivity extends AppCompatActivity {
                     public void run() {
                         for (Playlist mediastorePlaylist : arr){
                             // update the viewpager adapter with the mediastore playlists
-                            mainFragment.addPlaylist(mediastorePlaylist);
+                            mainFragmentPrimary.addPlaylist(mediastorePlaylist);
                         }
                     }
                 });
@@ -438,9 +356,7 @@ public class MainActivity extends AppCompatActivity {
     public void initMusicUI(){
         // init main ui
         initNotification();
-        initMainDisplay();
-        initSeekbar();
-        initSlidingUpPanel();
+        mainFragmentSecondary.initMainFragmentSecondaryUI();
     }
 
     @Override
@@ -457,363 +373,6 @@ public class MainActivity extends AppCompatActivity {
                 }
                 break;
         }
-    }
-
-    /**
-     * Initialize views and buttons in the sliding up panel
-     * The sliding up panel can be expanded upon click, but not collapsed
-     * Buttons will be invisible and unclickable depending on the panel state
-     */
-    public void initSlidingUpPanel() {
-        // init functionality for buttons on sliding menu
-        initSlidingUpPanelButtons();
-
-        // enable sliding menu text marquee left-right scrolling
-        slidingUp_songName.setSelected(true);
-        slidingUp_artistName.setSelected(true);
-
-        // init state of sliding panel views
-        if (slidingUpPanelLayout.getPanelState() == SlidingUpPanelLayout.PanelState.EXPANDED) {
-            // set layout alphas
-            slidingUpMenuLayout.setAlpha(0);
-            mainActivityRelativeLayout.setAlpha(1);
-
-            // disable buttons in sliding menu
-            slidingUp_prev_btn.setClickable(false);
-            slidingUp_pauseplay_btn.setClickable(false);
-            slidingUp_next_btn.setClickable(false);
-
-            // enable buttons in main display
-            mainDisplay_slidedown_btn.setClickable(true);
-        }
-        else if (slidingUpPanelLayout.getPanelState() == SlidingUpPanelLayout.PanelState.COLLAPSED) {
-            // set layout alphas
-            slidingUpMenuLayout.setAlpha(1);
-            mainActivityRelativeLayout.setAlpha(0);
-
-            // enable buttons in sliding menu
-            slidingUp_prev_btn.setClickable(true);
-            slidingUp_pauseplay_btn.setClickable(true);
-            slidingUp_next_btn.setClickable(true);
-
-            // disable buttons in main display
-            mainDisplay_slidedown_btn.setClickable(false);
-        }
-
-        // init slide listener
-        slidingUpPanelLayout.addPanelSlideListener(new SlidingUpPanelLayout.PanelSlideListener() {
-            @Override
-            public void onPanelSlide(View panel, float slideOffset) {
-                slidingUpMenuLayout.setAlpha(1 - slideOffset);
-                mainActivityRelativeLayout.setAlpha(slideOffset);
-            }
-
-            @Override
-            public void onPanelStateChanged(View panel, SlidingUpPanelLayout.PanelState previousState, SlidingUpPanelLayout.PanelState newState) {
-                if (slidingUpPanelLayout.getPanelState() == SlidingUpPanelLayout.PanelState.EXPANDED) {
-                    // disable buttons in sliding menu
-                    slidingUp_prev_btn.setClickable(false);
-                    slidingUp_pauseplay_btn.setClickable(false);
-                    slidingUp_next_btn.setClickable(false);
-
-                    // enable buttons in main display
-                    mainDisplay_slidedown_btn.setClickable(true);
-                }
-                else if (slidingUpPanelLayout.getPanelState() == SlidingUpPanelLayout.PanelState.COLLAPSED) {
-                    // enable buttons in sliding menu
-                    slidingUp_prev_btn.setClickable(true);
-                    slidingUp_pauseplay_btn.setClickable(true);
-                    slidingUp_next_btn.setClickable(true);
-
-                    // disable buttons in main display
-                    mainDisplay_slidedown_btn.setClickable(false);
-                }
-            }
-        });
-
-        // enables expand clicking only for the menu (and disables it for the entire relativelayout)
-        slidingUpPanelLayout.getChildAt(1).setOnClickListener(null);
-        slidingUpMenuLayout.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                if (slidingUpPanelLayout.getPanelState() == SlidingUpPanelLayout.PanelState.COLLAPSED) {
-                    slidingUpPanelLayout.setPanelState(SlidingUpPanelLayout.PanelState.EXPANDED);
-                }
-            }
-        });
-    }
-
-    /**
-     * Initializes the three main buttons in sliding up panel:
-     * Pauseplay, previous, and next
-     * Each one will trigger a unique event from MusicPlayerService
-     */
-    @SuppressLint("ClickableViewAccessibility")
-    @TargetApi(21)
-    public void initSlidingUpPanelButtons() {
-        Messenger slidingUpPanelMessenger = new Messenger(messageHandler);
-
-        // init pauseplay button and ripple drawable
-        slidingUp_pauseplay_btn_ripple = (RippleDrawable) slidingUp_pauseplay_btn.getBackground();
-
-        slidingUp_pauseplayIntent = new Intent(this, MusicPlayerService.class);
-        slidingUp_pauseplayIntent.putExtra("pauseplay", slidingUpPanelMessenger);
-
-        slidingUp_pauseplay_btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startService(slidingUp_pauseplayIntent);
-            }
-        });
-
-        // init next button and ripple drawable
-        slidingUp_next_btn_ripple = (RippleDrawable) slidingUp_next_btn.getBackground();
-
-        slidingUp_nextIntent = new Intent(this, MusicPlayerService.class);
-        slidingUp_nextIntent.putExtra("next", slidingUpPanelMessenger);
-
-        slidingUp_next_btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startService(slidingUp_nextIntent);
-            }
-        });
-
-        // init prev button and ripple drawable
-        slidingUp_prev_btn_ripple = (RippleDrawable) slidingUp_prev_btn.getBackground();
-
-        slidingUp_prevIntent = new Intent(this, MusicPlayerService.class);
-        slidingUp_prevIntent.putExtra("prev", slidingUpPanelMessenger);
-
-        slidingUp_prev_btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startService(slidingUp_prevIntent);
-            }
-        });
-    }
-
-    /**
-     * Initialize the gradient, which can be mutated based on the on the theme and album art
-     */
-    @TargetApi(16)
-    public void initMainGradient() {
-        // set up gradient background which can be mutated
-        mainDisplay_mainGradient = (GradientDrawable) ResourcesCompat.getDrawable(this.getResources(), R.drawable.gradient_default, null);
-        mainDisplay_mainGradient.mutate();
-        mainActivityRelativeLayout.setBackground(mainDisplay_mainGradient);
-    }
-
-    /**
-     * Initializes the main display details, including the album art, song name, and artist name
-     * On button click (overlaps with albumart), scales the view and button with an animation
-     */
-    public void initMainDisplay() {
-        // init album art corner radius animations
-        mainDisplay_albumArt_cardView_animator_round = ObjectAnimator
-                .ofFloat(mainDisplay_albumArt_cardView, "radius", (float) mainDisplay_albumArt_cardView.getWidth() / 2)
-                .setDuration(350);
-        mainDisplay_albumArt_cardView_animator_square = ObjectAnimator
-                .ofFloat(mainDisplay_albumArt_cardView, "radius", (float) mainDisplay_albumArt_cardView.getWidth() / 10)
-                .setDuration(350);
-
-        // if the intended album art height is greater than the actual height, resize based on screen dimensions
-        if (mainDisplay_albumArt_cardView.getLayoutParams().height > mainDisplay_albumArt_cardView.getHeight()) {
-            // out fields in [x,y] format
-            int[] coords1 = new int[2];
-            int[] coords2 = new int[2];
-            mainDisplay_playlistHeader.getLocationOnScreen(coords1);
-            mainDisplay_songTitle.getLocationOnScreen(coords2);
-
-            // resize album art to fit between the playlist header and song title
-            float height_position1 = (float) coords1[1] + (float) (mainDisplay_playlistHeader.getHeight() * 2);
-            float height_position2 = coords2[1] - (mainDisplay_songTitle.getHeight() * 2);
-            float height_difference = Math.abs(height_position1 - height_position2);
-            float shrink_factor = height_difference / mainDisplay_albumArt_cardView.getLayoutParams().height;
-
-            mainDisplay_albumArt_cardView.getLayoutParams().height *= shrink_factor;
-            mainDisplay_albumArt_cardView.getLayoutParams().width *= shrink_factor;
-        }
-
-        // init main display album art circular if different from xml (less rounded corners)
-        if (isAlbumArtCircular) {
-            mainDisplay_albumArt_cardView.setScaleX(0.95f);
-            mainDisplay_albumArt_cardView.setScaleY(0.95f);
-            mainDisplay_albumArt_cardView.setRadius((float) mainDisplay_albumArt_cardView.getWidth() / 2);
-        }
-
-        // init album art size toggle button
-        mainDisplay_albumArt_btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                toggleLargeAlbumArt();
-                databaseRepository.updateMetadataIsAlbumArtCircular(isAlbumArtCircular);
-            }
-        });
-
-        // enable playlist header, song title, and song artist text with marquee scrolling
-        mainDisplay_playlistHeader.setSelected(true);
-        mainDisplay_songTitle.setSelected(true);
-        mainDisplay_songArtist.setSelected(true);
-    }
-
-    /**
-     * Initializes the main buttons in main activity:
-     * Pauseplay, Previous, Next, Shuffle
-     * Pauseplay, Previous, and Next will trigger a unique event from MusicPlayerService
-     */
-    @TargetApi(23)
-    public void initMainButtons() {
-        Messenger mainMessenger = new Messenger(messageHandler);
-
-        // init pauseplay button click functionality and its ripple
-        mainDisplay_pauseplay_btn_ripple = (RippleDrawable) mainDisplay_pauseplay_btn.getBackground();
-        mainDisplay_pauseplay_btn_ripple.setRadius(70);
-        mainPausePlayIntent = new Intent(this, MusicPlayerService.class);
-        mainPausePlayIntent.putExtra("pauseplay", mainMessenger);
-
-        mainDisplay_pauseplay_btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startService(mainPausePlayIntent);
-            }
-        });
-
-        // init next button click functionality and its ripple
-        mainDisplay_next_btn_ripple = (RippleDrawable) mainDisplay_next_btn.getBackground();
-        mainDisplay_next_btn_ripple.setRadius(70);
-        mainNextIntent = new Intent(this, MusicPlayerService.class);
-        mainNextIntent.putExtra("next", mainMessenger);
-
-        mainDisplay_next_btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startService(mainNextIntent);
-            }
-        });
-
-        // init prev button click functionality and its ripple
-        mainDisplay_prev_btn_ripple = (RippleDrawable) mainDisplay_prev_btn.getBackground();
-        mainDisplay_prev_btn_ripple.setRadius(70);
-        mainPrevIntent = new Intent(this, MusicPlayerService.class);
-        mainPrevIntent.putExtra("prev", mainMessenger);
-
-        mainDisplay_prev_btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startService(mainPrevIntent);
-            }
-        });
-
-        // init slidedown button functionality and its ripple
-        mainDisplay_slidedown_btn_ripple = (RippleDrawable) mainDisplay_slidedown_btn.getBackground();
-        mainDisplay_slidedown_btn_ripple.setRadius(50);
-        mainDisplay_slidedown_btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                slidingUpPanelLayout.setPanelState(SlidingUpPanelLayout.PanelState.COLLAPSED);
-            }
-        });
-
-        // init shuffle button functionality and its ripple
-        mainDisplay_shuffle_btn_ripple = (RippleDrawable) mainDisplay_shuffle_btn.getBackground();
-        mainDisplay_shuffle_btn_ripple.setRadius(70);
-        mainDisplay_shuffle_btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                // toggle shuffle button
-                if (isShuffled){
-                    isShuffled = false;
-                    mainDisplay_shuffle_btn.setImageAlpha(40);
-                    current_playlist = current_playlist.unshufflePlaylist(random_seed);
-                }
-                else {
-                    isShuffled = true;
-                    random_seed = Math.abs(new Random().nextInt());
-                    mainDisplay_shuffle_btn.setImageAlpha(255);
-                    current_playlist = current_playlist.shufflePlaylist(random_seed);
-                }
-            }
-        });
-
-        // init repeat button functionality and its ripple
-        mainDisplay_repeat_btn_ripple = (RippleDrawable) mainDisplay_repeat_btn.getBackground();
-        mainDisplay_repeat_btn_ripple.setRadius(70);
-        mainDisplay_repeat_btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                int next_repeat_status = repeat_status + 1 > 2 ? 0 : repeat_status + 1;
-                toggleRepeatButton(next_repeat_status);
-                repeat_status = next_repeat_status;
-            }
-        });
-    }
-
-    /**
-     * Initializes the seekbar and the textviews for current position and max duration
-     * Time is converted from milliseconds to HH:MM:SS format
-     * Textview for current position is updated upon user changing the seekbar
-     */
-    public void initSeekbar() {
-        // init thread for seekbar updates
-        HandlerThread seekbarHandlerThread = new HandlerThread("SeekbarHandler");
-        seekbarHandlerThread.start();
-        seekbarHandler = new MessageHandler(seekbarHandlerThread.getLooper());
-
-        // set the seekbar & textview duration and sync with mediaplayer
-        Intent seekBarDurationIntent = new Intent(this, MusicPlayerService.class);
-        seekBar_progressIntent = new Intent(this, MusicPlayerService.class);
-        seekBar_seekIntent = new Intent(this, MusicPlayerService.class);
-        Messenger mainMessenger = new Messenger(messageHandler);
-        seekBarDurationIntent.putExtra("seekbarDuration", mainMessenger);
-        startService(seekBarDurationIntent);
-
-        mainDisplay_seekBar.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
-            @Override
-            public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
-                String time = SongHelper.convertTime(seekBar.getProgress());
-                mainDisplay_musicPosition.setText(time);
-            }
-
-            @Override
-            public void onStartTrackingTouch(SeekBar seekBar) {
-                seekBar_isTracking = true;
-            }
-
-            @Override
-            public void onStopTrackingTouch(SeekBar seekBar) {
-                seekBar_seekIntent.putExtra("seekbarSeek", seekBar.getProgress());
-                startService(seekBar_seekIntent);
-
-                // update the seekPosition value in the local metadata
-                databaseRepository.updateMetadataSeek(seekBar.getProgress());
-
-                seekBar_isTracking = false;
-            }
-        });
-    }
-
-    /**
-     * Initializes the button for displaying details about a song in a scrollview
-     * Upon clicking the button, the current song will be sent in the intent
-     */
-    public void initInfoButton() {
-        mainDisplay_info_btn_ripple = (RippleDrawable) mainDisplay_info_btn.getBackground();
-        mainDisplay_info_btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (!isInfoDisplaying) {
-                    isInfoDisplaying = true;
-
-                    MusicDetailsFragment musicDetailsFragment = MusicDetailsFragment.getInstance(current_song, fullSongMetadataHashMap.get(current_song.getId()));
-                    getSupportFragmentManager().beginTransaction()
-                            .setReorderingAllowed(true)
-                            .add(R.id.fragment_playlist, musicDetailsFragment)
-                            .addToBackStack("musicDetailsFragment")
-                            .commit();
-                }
-            }
-        });
     }
 
     @TargetApi(19)
@@ -863,100 +422,11 @@ public class MainActivity extends AppCompatActivity {
      */
     public void updateTheme(int theme_resid) {
         setTheme(theme_resid);
-        mainFragment.updateFragmentColors();
-        updateSeekBarColors();
-        updateMainColors();
-        updateSlidingMenuColors();
+        mainFragmentPrimary.updateFragmentColors();
+        mainFragmentSecondary.updateFragmentColors();
     }
 
-    /**
-     * updates the two main gradients and the text colors in the main display
-     * gradient colors are derived from the current theme and song
-     */
-    @TargetApi(21)
-    public void updateMainColors() {
-        int textSongColor = ThemeColors.getColor(ThemeColors.TITLE_TEXT_COLOR);
-        int textArtistColor = ThemeColors.getColor(ThemeColors.TITLE_TEXT_COLOR);
-        int textSeekbarColor = ThemeColors.getColor(ThemeColors.TITLE_TEXT_COLOR);
-        int primaryColor = ThemeColors.getColor(ThemeColors.COLOR_PRIMARY);
-        int secondaryColor= ThemeColors.getGradientColor();
 
-        mainDisplay_songTitle.setTextColor(textSongColor);
-        mainDisplay_songArtist.setTextColor(textArtistColor);
-        mainDisplay_musicPosition.setTextColor(textSeekbarColor);
-        mainDisplay_musicDuration.setTextColor(textSeekbarColor);
-        mainDisplay_playlistHeader.setTextColor(textSongColor);
-        mainDisplay_mainGradient.setColors(new int[]{primaryColor, secondaryColor});
-        mainDisplay_mainGradient.setOrientation(GradientDrawable.Orientation.BL_TR);
-
-        // update drawable vector colors
-        Drawable unwrappedDrawableSlideDown = mainDisplay_slidedown_btn.getDrawable();
-        Drawable unwrappedDrawableInfo = mainDisplay_info_btn.getDrawable();
-        Drawable unwrappedDrawablePauseplay = mainDisplay_pauseplay_btn.getDrawable();
-        Drawable unwrappedDrawableNext = mainDisplay_next_btn.getDrawable();
-        Drawable unwrappedDrawablePrev = mainDisplay_prev_btn.getDrawable();
-        Drawable unwrappedDrawableShuffle = mainDisplay_shuffle_btn.getDrawable();
-        Drawable unwrappedDrawableRepeat = mainDisplay_repeat_btn.getDrawable();
-        Drawable wrappedDrawableSlideDown = DrawableCompat.wrap(unwrappedDrawableSlideDown);
-        Drawable wrappedDrawableInfo = DrawableCompat.wrap(unwrappedDrawableInfo);
-        Drawable wrappedDrawablePauseplay = DrawableCompat.wrap(unwrappedDrawablePauseplay);
-        Drawable wrappedDrawableNext = DrawableCompat.wrap(unwrappedDrawableNext);
-        Drawable wrappedDrawablePrev = DrawableCompat.wrap(unwrappedDrawablePrev);
-        Drawable wrappedDrawableShuffle = DrawableCompat.wrap(unwrappedDrawableShuffle);
-        Drawable wrappedDrawableRepeat = DrawableCompat.wrap(unwrappedDrawableRepeat);
-        DrawableCompat.setTint(wrappedDrawableSlideDown, getResources().getColor(ThemeColors.getMainDrawableVectorColorId()));
-        DrawableCompat.setTint(wrappedDrawableInfo, getResources().getColor(ThemeColors.getMainDrawableVectorColorId()));
-        DrawableCompat.setTint(wrappedDrawablePauseplay, getResources().getColor(ThemeColors.getMainDrawableVectorColorId()));
-        DrawableCompat.setTint(wrappedDrawableNext, getResources().getColor(ThemeColors.getMainDrawableVectorColorId()));
-        DrawableCompat.setTint(wrappedDrawablePrev, getResources().getColor(ThemeColors.getMainDrawableVectorColorId()));
-        DrawableCompat.setTint(wrappedDrawableShuffle, getResources().getColor(ThemeColors.getMainDrawableVectorColorId()));
-        DrawableCompat.setTint(wrappedDrawableRepeat, getResources().getColor(ThemeColors.getMainDrawableVectorColorId()));
-
-        // update ripple colors
-        mainDisplay_slidedown_btn_ripple.setColor(ColorStateList.valueOf(getResources().getColor(ThemeColors.getMainRippleDrawableColorId())));
-        mainDisplay_info_btn_ripple.setColor(ColorStateList.valueOf(getResources().getColor(ThemeColors.getMainRippleDrawableColorId())));
-        mainDisplay_pauseplay_btn_ripple.setColor(ColorStateList.valueOf(getResources().getColor(ThemeColors.getMainRippleDrawableColorId())));
-        mainDisplay_next_btn_ripple.setColor(ColorStateList.valueOf(getResources().getColor(ThemeColors.getMainRippleDrawableColorId())));
-        mainDisplay_prev_btn_ripple.setColor(ColorStateList.valueOf(getResources().getColor(ThemeColors.getMainRippleDrawableColorId())));
-        mainDisplay_shuffle_btn_ripple.setColor(ColorStateList.valueOf(getResources().getColor(ThemeColors.getMainRippleDrawableColorId())));
-        mainDisplay_repeat_btn_ripple.setColor(ColorStateList.valueOf(getResources().getColor(ThemeColors.getMainRippleDrawableColorId())));
-    }
-
-    /**
-     * updates the background, text, and button colors of the sliding menu
-     * text and button colors are decided as the swatch that contrasts the most with the background
-     */
-    @TargetApi(21)
-    public void updateSlidingMenuColors() {
-        int base = ThemeColors.getColor(ThemeColors.COLOR_SECONDARY);
-        int contrastColor = ThemeColors.getContrastColor(base);
-
-        // change color of sliding menu, its buttons, and text
-        slidingUp_songName.setTextColor(contrastColor);
-        slidingUp_artistName.setTextColor(contrastColor);
-        slidingUpMenuLayout.setBackgroundColor(base);
-        Drawable unwrappedDrawablePauseplay = slidingUp_pauseplay_btn.getDrawable();
-        Drawable unwrappedDrawableNext = slidingUp_next_btn.getDrawable();
-        Drawable unwrappedDrawablePrev = slidingUp_prev_btn.getDrawable();
-        Drawable wrappedDrawablePauseplay = DrawableCompat.wrap(unwrappedDrawablePauseplay);
-        Drawable wrappedDrawableNext = DrawableCompat.wrap(unwrappedDrawableNext);
-        Drawable wrappedDrawablePrev = DrawableCompat.wrap(unwrappedDrawablePrev);
-        DrawableCompat.setTint(wrappedDrawablePauseplay, contrastColor);
-        DrawableCompat.setTint(wrappedDrawableNext, contrastColor);
-        DrawableCompat.setTint(wrappedDrawablePrev, contrastColor);
-
-        // change color of the button ripples
-        slidingUp_pauseplay_btn_ripple.setColor(ColorStateList.valueOf(contrastColor));
-        slidingUp_next_btn_ripple.setColor(ColorStateList.valueOf(contrastColor));
-        slidingUp_prev_btn_ripple.setColor(ColorStateList.valueOf(contrastColor));
-    }
-
-    @TargetApi(23)
-    private void updateSeekBarColors(){
-        mainDisplay_seekBar.getThumb().setTint(ThemeColors.getColor(ThemeColors.TITLE_TEXT_COLOR));
-        mainDisplay_seekBar.getProgressDrawable().setTint(ThemeColors.getColor(ThemeColors.TITLE_TEXT_COLOR));
-        ((RippleDrawable) mainDisplay_seekBar.getBackground()).setColor(ColorStateList.valueOf(getResources().getColor(ThemeColors.getRippleDrawableColorId())));
-    }
 
     /**
      * Cleans playlist database of any songs that can no longer be found in the full list of songs
@@ -1008,10 +478,10 @@ public class MainActivity extends AppCompatActivity {
         }
 
         // set shuffle button transparency using the isShuffled metadata
-        mainDisplay_shuffle_btn.setImageAlpha(isShuffled ? 255 : 40);
+        mainFragmentSecondary.setShuffleBtnAlpha(isShuffled ? 255 : 40);
 
         // set repeat button appearance using repeatStatus metadata
-        toggleRepeatButton(repeat_status);
+        mainFragmentSecondary.toggleRepeatButton(repeat_status);
 
         // retrieve all playlists that exist in mediastore if this hasn't been done before
         if (!isMediaStorePlaylistsImported){
@@ -1033,9 +503,7 @@ public class MainActivity extends AppCompatActivity {
             initMusicUI();
 
             // inform the music service about the seekbar's position from the metadata
-            seekBar_seekIntent.putExtra("seekbarSeek", seekPosition);
-            startService(seekBar_seekIntent);
-            mainDisplay_seekBar.setProgress(seekPosition);
+            mainFragmentSecondary.updateSeekbarProgress(seekPosition);
         }
 
         // set the current theme and generate theme values
@@ -1046,57 +514,6 @@ public class MainActivity extends AppCompatActivity {
         updateTheme(themeResourceId);
         SongListTab.setScrollSelection(songtab_scrollindex, songtab_scrolloffset);
         isMetadataLoaded = true;
-    }
-
-    /**
-     * Helper method to enable scaling of album art and visibility of song name & artist
-     * Based on largeAlbumArt boolean conditions:
-     * false - album art is not large and song name & artist are visible
-     * true - album art is large and hides song name & artist
-     */
-    @TargetApi(16)
-    public void toggleLargeAlbumArt(){
-        if (isAlbumArtCircular) {
-            isAlbumArtCircular = false;
-            mainDisplay_albumArt_cardView.animate().scaleX(1f).scaleY(1f);
-            mainDisplay_albumArt_cardView_animator_square.start();
-        }
-        else{
-            isAlbumArtCircular = true;
-            mainDisplay_albumArt_cardView.animate().scaleX(0.95f).scaleY(0.95f);
-            mainDisplay_albumArt_cardView_animator_round.start();
-        }
-    }
-
-    /**
-     * Helper method to set the appearance of the repeat button, based on the repeat status
-     * @param repeatStatus 0 for disable, 1 for repeat playlist, 2 for repeat one song
-     */
-    public void toggleRepeatButton(int repeatStatus){
-        switch (repeatStatus){
-            // disable repeat
-            case 0:
-                mainDisplay_repeat_btn.setImageResource(R.drawable.ic_repeat28dp);
-                mainDisplay_repeat_btn.setImageAlpha(40);
-                break;
-
-            // repeat playlist
-            case 1:
-                mainDisplay_repeat_btn.setImageResource(R.drawable.ic_repeat28dp);
-                mainDisplay_repeat_btn.setImageAlpha(255);
-                break;
-
-            // repeat one song
-            case 2:
-                mainDisplay_repeat_btn.setImageResource(R.drawable.ic_repeat_one28dp);
-                mainDisplay_repeat_btn.setImageAlpha(255);
-                break;
-        }
-
-        // set appropriate colors for the button
-        Drawable unwrappedDrawableRepeat = mainDisplay_repeat_btn.getDrawable();
-        Drawable wrappedDrawableRepeat = DrawableCompat.wrap(unwrappedDrawableRepeat);
-        DrawableCompat.setTint(wrappedDrawableRepeat, getResources().getColor(ThemeColors.getMainDrawableVectorColorId()));
     }
 
     /**
@@ -1115,7 +532,7 @@ public class MainActivity extends AppCompatActivity {
                     cleanPlaylistDatabase();
                     SongListAdapter songListadapter = new SongListAdapter(this, R.layout.adapter_song_layout, fullSongList, this);
                     PlaylistAdapter playlistAdapter = new PlaylistAdapter(this, R.layout.adapter_playlist_layout, playlistList, this);
-                    mainFragment.setAdapters(songListadapter, playlistAdapter);
+                    mainFragmentPrimary.setAdapters(songListadapter, playlistAdapter);
 
                     // initialize current playlist from database, if possible
                     databaseRepository.asyncGetCurrentPlaylist();
@@ -1137,13 +554,13 @@ public class MainActivity extends AppCompatActivity {
                     setupMetadata();
                     break;
                 case DatabaseRepository.ASYNC_INSERT_PLAYLIST:
-                    mainFragment.updateMainFragment(object, DatabaseRepository.ASYNC_INSERT_PLAYLIST);
+                    mainFragmentPrimary.updateMainFragment(object, DatabaseRepository.ASYNC_INSERT_PLAYLIST);
                     break;
                 case DatabaseRepository.ASYNC_MODIFY_PLAYLIST:
-                    mainFragment.updateMainFragment(object, DatabaseRepository.ASYNC_MODIFY_PLAYLIST);
+                    mainFragmentPrimary.updateMainFragment(object, DatabaseRepository.ASYNC_MODIFY_PLAYLIST);
                     break;
                 case DatabaseRepository.ASYNC_DELETE_PLAYLISTS_BY_ID:
-                    mainFragment.updateMainFragment(object, DatabaseRepository.ASYNC_DELETE_PLAYLISTS_BY_ID);
+                    mainFragmentPrimary.updateMainFragment(object, DatabaseRepository.ASYNC_DELETE_PLAYLISTS_BY_ID);
                     break;
                 case DatabaseRepository.ASYNC_GET_METADATA:
                     // metadata object guaranteed not null
@@ -1176,7 +593,25 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public View getMainActivityLayout(){
-        return slidingUpPanelLayout;
+        return mainActivityLayout;
+    }
+    public DatabaseRepository getDatabaseRepository(){
+        return databaseRepository;
+    }
+    public HashMap<Integer, SongMetadata> getFullSongMetadataHashMap(){
+        return fullSongMetadataHashMap;
+    }
+    public boolean getIsAlbumArtCircular(){
+        return isAlbumArtCircular;
+    }
+    public boolean getIsInfoDisplaying(){
+        return isInfoDisplaying;
+    }
+    public static int getRandom_seed(){
+        return random_seed;
+    }
+    public static boolean getIsShuffled(){
+        return isShuffled;
     }
     public static int getRepeat_status(){
         return repeat_status;
@@ -1197,6 +632,21 @@ public class MainActivity extends AppCompatActivity {
         return notificationChannel1;
     }
 
+    public void setIsAlbumArtCircular(boolean isAlbumArtCircular){
+        this.isAlbumArtCircular = isAlbumArtCircular;
+    }
+    public void setIsInfoDisplaying(boolean isInfoDisplaying){
+        this.isInfoDisplaying = isInfoDisplaying;
+    }
+    public static void setRandom_seed(int seed){
+        random_seed = seed;
+    }
+    public static void setIsShuffled(boolean shuffled){
+        isShuffled = shuffled;
+    }
+    public static void setRepeat_status(int status){
+        repeat_status = status;
+    }
     public static void setCurrent_song(Song song){
         current_song = song;
     }
@@ -1211,9 +661,6 @@ public class MainActivity extends AppCompatActivity {
     public static void setCurrent_transientPlaylist(Playlist transient_playlist){
         Playlist current_playlist = new Playlist(transient_playlist.getName(), transient_playlist.getSongList());
         setCurrent_playlist(current_playlist);
-    }
-    public void setIsInfoDisplaying(boolean isInfoDisplaying){
-        this.isInfoDisplaying = isInfoDisplaying;
     }
 
     /**
@@ -1239,18 +686,7 @@ public class MainActivity extends AppCompatActivity {
                     runOnUiThread(new Runnable() {
                         @Override
                         public void run() {
-                            mainDisplay_pauseplay_btn.setImageResource(R.drawable.ic_play28dp);
-                            slidingUp_pauseplay_btn.setImageResource(R.drawable.ic_play24dp);
-
-                            // change sliding menu pauseplay button color
-                            Drawable unwrappedDrawablePauseplay = slidingUp_pauseplay_btn.getDrawable();
-                            Drawable wrappedDrawablePauseplay = DrawableCompat.wrap(unwrappedDrawablePauseplay);
-                            DrawableCompat.setTint(wrappedDrawablePauseplay, ThemeColors.getContrastColor());
-
-                            // change main pauseplay button color
-                            Drawable unwrappedMainDrawablePauseplay = mainDisplay_pauseplay_btn.getDrawable();
-                            Drawable wrappedMainDrawablePauseplay = DrawableCompat.wrap(unwrappedMainDrawablePauseplay);
-                            DrawableCompat.setTint(wrappedMainDrawablePauseplay, getResources().getColor(ThemeColors.getMainDrawableVectorColorId()));
+                            mainFragmentSecondary.setPausePlayBtns(R.drawable.ic_play28dp, R.drawable.ic_play24dp);
                         }
                     });
                     notificationBuilder.mActions.set(1, new NotificationCompat.Action(R.drawable.ic_play24dp, "play", PendingIntent.getService(getApplicationContext(), 1, notificationPauseplayIntent, PendingIntent.FLAG_UPDATE_CURRENT)));
@@ -1267,7 +703,7 @@ public class MainActivity extends AppCompatActivity {
                         }
                         // use the seek position from the seekbar view
                         else {
-                            databaseRepository.updateMetadataSeek(mainDisplay_seekBar.getProgress());
+                            databaseRepository.updateMetadataSeek(mainFragmentSecondary.getSeekbarProgress());
                         }
                     }
                     break;
@@ -1275,18 +711,7 @@ public class MainActivity extends AppCompatActivity {
                     runOnUiThread(new Runnable() {
                         @Override
                         public void run() {
-                            mainDisplay_pauseplay_btn.setImageResource(R.drawable.ic_pause28dp);
-                            slidingUp_pauseplay_btn.setImageResource(R.drawable.ic_pause24dp);
-
-                            // change sliding menu pauseplay button color
-                            Drawable unwrappedDrawablePauseplay = slidingUp_pauseplay_btn.getDrawable();
-                            Drawable wrappedDrawablePauseplay = DrawableCompat.wrap(unwrappedDrawablePauseplay);
-                            DrawableCompat.setTint(wrappedDrawablePauseplay, ThemeColors.getContrastColor());
-
-                            // change main pauseplay button color
-                            Drawable unwrappedMainDrawablePauseplay = mainDisplay_pauseplay_btn.getDrawable();
-                            Drawable wrappedMainDrawablePauseplay = DrawableCompat.wrap(unwrappedMainDrawablePauseplay);
-                            DrawableCompat.setTint(wrappedMainDrawablePauseplay, getResources().getColor(ThemeColors.getMainDrawableVectorColorId()));
+                            mainFragmentSecondary.setPausePlayBtns(R.drawable.ic_pause28dp, R.drawable.ic_pause24dp);
                         }
                     });
                     notificationBuilder.mActions.set(1, new NotificationCompat.Action(R.drawable.ic_pause24dp, "pause", PendingIntent.getService(getApplicationContext(), 1, notificationPauseplayIntent, PendingIntent.FLAG_UPDATE_CURRENT)));
@@ -1301,13 +726,11 @@ public class MainActivity extends AppCompatActivity {
                 case MusicPlayerService.UPDATE_SEEKBAR_DURATION:
                     // init the seekbar & textview max duration and begin thread to track progress
                     final int musicMaxDuration = (int) bundle.get("time");
-                    final String time = SongHelper.convertTime(musicMaxDuration);
 
                     runOnUiThread(new Runnable() {
                         @Override
                         public void run() {
-                            mainDisplay_seekBar.setMax(musicMaxDuration);
-                            mainDisplay_musicDuration.setText(time);
+                            mainFragmentSecondary.setSeekbarMaxDuration(musicMaxDuration);
                         }
                     });
 
@@ -1318,7 +741,7 @@ public class MainActivity extends AppCompatActivity {
                         @Override
                         public void run() {
                             while (!isDestroyed) {
-                                if (!seekBar_isTracking) {
+                                if (!mainFragmentSecondary.getSeekbar_isTracking()) {
                                     startService(seekBar_progressIntent);
                                 }
                                 try {
@@ -1334,7 +757,7 @@ public class MainActivity extends AppCompatActivity {
                     break;
                 case MusicPlayerService.UPDATE_SEEKBAR_PROGRESS:
                     int musicCurrentPosition = (int) bundle.get("time");
-                    mainDisplay_seekBar.setProgress(musicCurrentPosition);
+                    mainFragmentSecondary.setSeekbarProgress(musicCurrentPosition);
                     break;
                 case MusicPlayerService.UPDATE_SONG:
                     // update main activity with the selected song from music list
@@ -1356,24 +779,12 @@ public class MainActivity extends AppCompatActivity {
                         e.printStackTrace();
                     }
 
-                    final int songDuration = current_song.getDuration();
 
                     // view changes must be done on the main ui thread
                     runOnUiThread(new Runnable() {
                         @Override
                         public void run() {
-                            // update sliding menu details
-                            slidingUp_songName.setText(current_song.getTitle());
-                            slidingUp_artistName.setText(current_song.getArtist());
-                            slidingUp_albumArt.setImageBitmap(current_albumImage);
-
-                            // update main activity details
-                            mainDisplay_playlistHeader.setText(current_playlist.getName());
-                            mainDisplay_songTitle.setText(current_song.getTitle());
-                            mainDisplay_songArtist.setText(current_song.getArtist());
-                            mainDisplay_albumArt.setImageBitmap(current_albumImage);
-                            mainDisplay_seekBar.setMax(songDuration);
-                            mainDisplay_musicDuration.setText(SongHelper.convertTime(songDuration));
+                            mainFragmentSecondary.updateMainSongDetails(current_song, current_playlist, current_albumImage);
                         }
                     });
 
@@ -1391,8 +802,7 @@ public class MainActivity extends AppCompatActivity {
                     runOnUiThread(new Runnable() {
                         @Override
                         public void run() {
-                            updateMainColors();
-                            updateSlidingMenuColors();
+                            mainFragmentSecondary.updateFragmentColors();
                         }
                     });
 
@@ -1408,7 +818,7 @@ public class MainActivity extends AppCompatActivity {
                         @Override
                         public void run() {
                             updateTheme(theme_resid);
-                            mainFragment.updateFragmentColors();
+                            mainFragmentPrimary.updateFragmentColors();
                         }
                     });
                     break;
@@ -1416,7 +826,7 @@ public class MainActivity extends AppCompatActivity {
                     runOnUiThread(new Runnable() {
                         @Override
                         public void run() {
-                            mainFragment.rotateThemeButton();
+                            mainFragmentPrimary.rotateThemeButton();
                         }
                     });
                     break;
